# shared/contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "LinkedIn Post Bot API",
    "description": "\n    AI-powered LinkedIn content automation API.\n    \n    ## Features\n    - Generate LinkedIn posts from GitHub activity\n    - Multiple AI writing templates (Standard, Build in Public, Thought Leadership, Job Search)\n    - Post history and analytics\n    - Image selection from Unsplash\n    - Scheduled posting\n    \n    ## OpenAPI Documentation\n    - OpenAPI JSON: `/openapi.json`\n    - Interactive Docs (Swagger): `/docs`  \n    - ReDoc: `/redoc`\n    ",
    "contact": {
      "name": "LinkedIn Post Bot",
      "url": "https://github.com/cliff-de-tech/Linkedin-Post-Bot"
    },
    "license": {
      "name": "MIT"
    },
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/checkout": {
      "post": {
        "tags": [
          "Payments"
        ],
        "summary": "Create Checkout",
        "description": "Initialize a Paystack hosted checkout flow to start a subscription.\n\nThe user will be redirected to Paystack's hosted checkout page.\nAfter payment, they'll be redirected to success_url or cancel_url.\n\n**Authentication Required**: Yes (Clerk JWT)\n\n**Request Body**:\n- `user_id`: Clerk user ID\n- `price_id`: Paystack plan code (from your Paystack dashboard)\n- `email`: Customer email (required by Paystack)\n- `success_url`: Where to redirect after successful payment\n- `cancel_url`: Where to redirect if user cancels\n\n**Response**:\n- `session_id`: Paystack transaction reference\n- `checkout_url`: Direct URL to the checkout page",
        "operationId": "create_checkout_api_checkout_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CheckoutRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CheckoutResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/billing-portal": {
      "post": {
        "tags": [
          "Payments"
        ],
        "summary": "Create Portal",
        "description": "Return a subscription management URL if your app provides one.\n\nPaystack does not provide a built-in hosted billing portal. If your app\nconfigures a custom management page, this endpoint returns that URL.\n\n**Authentication Required**: Yes (Clerk JWT)\n\n**Request Body**:\n- `user_id`: Clerk user ID\n- `return_url`: Where to redirect after management flow\n\n**Response**:\n- `portal_url`: URL to redirect user to",
        "operationId": "create_portal_api_billing_portal_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BillingPortalRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BillingPortalResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/subscription": {
      "post": {
        "tags": [
          "Payments"
        ],
        "summary": "Get Subscription",
        "description": "Get current subscription status for a user.\n\n**Authentication Required**: Yes (Clerk JWT)\n\n**Request Body**:\n- `user_id`: Clerk user ID\n\n**Response**:\n- `has_subscription`: Whether user has any subscription record\n- `status`: Subscription status (active, non_renewing, attention, etc.)\n- `plan_id`: Paystack plan code\n- `current_period_end`: Unix timestamp of current billing period end\n- `cancel_at_period_end`: Whether subscription is scheduled to cancel",
        "operationId": "get_subscription_api_subscription_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubscriptionStatusRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubscriptionStatusResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
    "/webhook/paystack": {
      "post": {
        "tags": [
          "Webhooks"
        ],
        "summary": "Paystack Webhook",
        "description": "Handle Paystack webhook events.\n\n**Authentication**: None (signature verified internally)\n\nThis endpoint receives events from Paystack such as:\n- `charge.success`\n- `subscription.create`\n- `subscription.not_renew`\n- `subscription.disable`\n- `invoice.payment_failed`\n\n**SECURITY**: Webhook signature is verified using PAYSTACK_WEBHOOK_SECRET\n(or PAYSTACK_SECRET_KEY if no explicit webhook secret is configured).\nNever process events without verification.",
        "operationId": "paystack_webhook_webhook_paystack_post",
        "parameters": [
          {
            "name": "x-paystack-signature",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "title": "X-Paystack-Signature"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BillingPortalRequest": {
        "properties": {
          "user_id": {
            "type": "string",
            "maxLength": 64,
            "minLength": 1,
            "title": "User Id"
          },
          "return_url": {
            "type": "string",
            "maxLength": 500,
            "title": "Return Url",
            "default": "http://localhost:3000/settings"
          }
        },
        "type": "object",
        "required": [
          "user_id"
        ],
        "title": "BillingPortalRequest",
        "description": "Request to open a subscription management page."
      },
      "BillingPortalResponse": {
        "properties": {
          "portal_url": {
            "type": "string",
            "title": "Portal Url"
          }
        },
        "type": "object",
        "required": [
          "portal_url"
        ],
        "title": "BillingPortalResponse",
        "description": "Response with subscription-management URL."
      },
      "CheckoutRequest": {
        "properties": {
          "user_id": {
            "type": "string",
            "maxLength": 64,
            "minLength": 1,
            "title": "User Id"
          },
          "price_id": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "title": "Price Id",
            "description": "Paystack plan code (PLN_xxxxx)"
          },
          "email": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 254
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "success_url": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 500
              },
              {
                "type": "null"
              }
            ],
            "title": "Success Url"
          },
          "cancel_url": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 500
              },
              {
                "type": "null"
              }
            ],
            "title": "Cancel Url"
          }
        },
        "type": "object",
        "required": [
          "user_id",
          "price_id"
        ],
        "title": "CheckoutRequest",
        "description": "Request to create a checkout session."
      },
      "CheckoutResponse": {
        "properties": {
          "session_id": {
            "type": "string",
            "title": "Session Id"
          },
          "checkout_url": {
            "type": "string",
            "title": "Checkout Url"
          }
        },
        "type": "object",
        "required": [
          "session_id",
          "checkout_url"
        ],
        "title": "CheckoutResponse",
        "description": "Response with checkout session details."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "SubscriptionStatusRequest": {
        "properties": {
          "user_id": {
            "type": "string",
            "maxLength": 64,
            "minLength": 1,
            "title": "User Id"
          }
        },
        "type": "object",
        "required": [
          "user_id"
        ],
        "title": "SubscriptionStatusRequest",
        "description": "Request to get subscription status."
      },
      "SubscriptionStatusResponse": {
        "properties": {
          "has_subscription": {
            "type": "boolean",
            "title": "Has Subscription"
          },
          "status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status"
          },
          "plan_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Plan Id"
          },
          "current_period_end": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Period End"
          },
          "cancel_at_period_end": {
            "type": "boolean",
            "title": "Cancel At Period End",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "has_subscription"
        ],
        "title": "SubscriptionStatusResponse",
        "description": "Response with subscription details."
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}